fix: plumb full 64-bit ReadoutConfig::seed into HCNN weight init
ReadoutConfig::seed was unsigned, so 64-bit assignments (e.g. NARMA
7934791766227647176) truncated to low-32 (816937672) before HCNN init.
Promote seed/weight_seed to uint64_t end-to-end; keep mt19937(seed32)
when high half is zero for historical bit-identical campaigns.

+++ python/bindings.cpp
@@ -33,7 +33,7 @@ PYBIND11_MODULE(_core, m)
                          float readout_lr_max, float readout_lr_min_frac,
                          int readout_lr_decay_epochs, float readout_weight_decay,
                          float readout_momentum, const char* readout_activation,
-                         unsigned readout_seed,
+                         uint64_t readout_seed,
                          size_t readout_num_threads,
                          bool readout_restore_best_epoch,
                          float readout_best_epoch_holdout_frac) {
@@ -101,7 +101,7 @@ PYBIND11_MODULE(_core, m)
             py::arg("readout_weight_decay")     = 0.0f,
             py::arg("readout_momentum")         = 0.0f,
             py::arg("readout_activation")       = "tanh",
-            py::arg("readout_seed")             = 42u,
+            py::arg("readout_seed")             = 42ULL,
             py::arg("readout_num_threads")      = 0ULL,
             py::arg("readout_restore_best_epoch") = true,
             py::arg("readout_best_epoch_holdout_frac") = 0.0f)
